Version capable of handling different action space
The games in the Gymnasium library are limited to two types of action spaces, meaning there are two distinct classes for implementing the tasks. These classes have different methods, one of which pertains to acquiring the action space. That being said, this version is capable of handling both action space scenarios, thereby generalizing the solution for both discrete and continuous action spaces.