Updated to Handle JSON
This code has been updated to handle JSON, meaning that the Args class will receive a JSON file containing all the configuration settings for the model implementation. In addition to this main change, attributes related to controlling the layers of the actor networks and the critic network have also been added.