Second Implementation (Without Tensorboard)
The main ideia here is to create class wich can use like others librarys in python. Only essential dependencies were kept. A dictionary was added to store the main metrics of the model, and a method was added to the PPOTrainer class to visualize these metrics. Additionally, a function was created (which is not in the class module but can be added) to visualize the results of the policy learned by the agent.